+++ js/src/styles/_theme.scss
@@ -411,70 +411,3 @@ textarea.editor {
 .pagination {
   user-select: none;
 }
-
-//$item-width: 100px;
-//$item-height: 50px;
-//$file-summary-height: $item-height + 40px;
-//.dropzone {
-//  position: relative;
-//  outline: none;
-//  .previews {
-//    position: relative;
-//    display: flex;
-//    flex-wrap: wrap;
-//    margin-top: 5px;
-//  }
-//  .file-summary {
-//    font-size: 12px;
-//    margin: 0 20px;
-//    text-align: center;
-//    height: $file-summary-height;
-//    //border: 1px dotted #888;
-//  }
-//  .file-summary-main {
-//    height: 100%;
-//    width: $item-width;
-//    white-space: nowrap;
-//    text-overflow: ellipsis;
-//    overflow: hidden;
-//  }
-//  .file-preview {
-//    margin: auto;
-//    display: flex;
-//    align-items: center;
-//    justify-content: center;
-//    width: $item-width;
-//    height: $item-height;
-//    img {
-//      width: $item-width;
-//      max-height: $item-height;
-//      object-fit: cover;
-//    }
-//  }
-//  .preview-overlay {
-//    align-items: center;
-//    justify-content: center;
-//    position: relative;
-//    top: -$file-summary-height;
-//    width: 100%;
-//    height: $file-summary-height;
-//    z-index: 999;
-//    color: #888;
-//    background: rgba(255, 255, 255, 0.5);
-//    cursor: pointer;
-//  }
-//  .full-overlay {
-//    align-items: center;
-//    justify-content: center;
-//    position: absolute;
-//    width: 100%;
-//    height: 100%;
-//    top: 0;
-//    left: 0;
-//    z-index: 999;
-//    border: 4px dashed #ccc;
-//    color: #ccc;
-//    border-radius: 4px;
-//    background: rgba(255, 255, 255, 0.8);
-//  }
-//}
